backend/requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
PyJWT
sqlparse
psycopg2-binary
python-dotenv